+++ R/hyp_to_table.R
@@ -4,7 +4,7 @@
 #' @param file_path A file path for hyp objects and directory for multihyp objects
 #' @param sep The field separator string
 #' @param cols Dataframe columns to include
-#' @param version Add header with versioning information
+#' @param versioning Add header with versioning information
 #' @return NULL
 #'
 #' @examples
@@ -19,8 +19,9 @@
 #' hyp_to_table(hyp_obj, file_path="pathways.txt")
 #'
 #' @importFrom magrittr %>% extract
+#' 
 #' @export
-hyp_to_table <- function(hyp_obj, file_path, sep="\t", cols=NULL, version=TRUE) {
+hyp_to_table <- function(hyp_obj, file_path, sep="\t", cols=NULL, versioning=TRUE) {
 
     stopifnot(is(hyp_obj, "hyp") | is(hyp_obj, "multihyp"))
 
@@ -55,17 +56,15 @@ hyp_to_table <- function(hyp_obj, file_path, sep="\t", cols=NULL, version=TRUE) 
             cols <- seq_len(ncol(df))
         }
 
-        if (version) {
+        if (versioning) {
 
-            name <- hyp_obj$args$genesets$name
-            version <- hyp_obj$args$genesets$version
-            header <- list(.format_str("# Generated with hypeR v{1}", packageVersion("hypeR")),
-                           .format_str("# Using the following genesets: {1} {2}", name, version))
-
+            header <- matrix(c(names(hyp_obj$info), as.character(hyp_obj$info)), ncol=2)
+            header[,1] <- paste("#", header[,1], sep="")
+            
             write.table(x=header, 
                         file=file_path, 
-                        quote=FALSE, 
-                        sep="\n",
+                        quote=FALSE,
+                        sep=sep,
                         col.names=FALSE, 
                         row.names=FALSE,
                         append=FALSE)
@@ -76,11 +75,11 @@ hyp_to_table <- function(hyp_obj, file_path, sep="\t", cols=NULL, version=TRUE) 
         }
 
         suppressWarnings(write.table(x=df[,cols,drop=FALSE],
-                        file=file_path,
-                        quote=FALSE,
-                        sep=sep,
-                        col.names=TRUE,
-                        row.names=FALSE,
-                        append=append))
+                                     file=file_path,
+                                     quote=FALSE,
+                                     sep=sep,
+                                     col.names=TRUE,
+                                     row.names=FALSE,
+                                     append=append))
     }
 }