+++ R/pvector.R
@@ -1,43 +1,35 @@
 #' A push/pop capable vector
 #'
-#' @name pvector
-#'
-#' @section Arguments:
-#' \describe{
-#'   \item{values}{A vector of values}
-#' }
-#'
-#' @section Methods:
-#' 
-#' \code{print(pvector)} prints vector values.
-#' 
-#' \code{pvector$length()} returns length of vector values.
-#'  
-#' \code{pvector$pop()} pops vector, returning popped value.
-#'
-#' \code{pvector$push()} pushes to vector.
-#'
-#' @examples
-#' pv <- pvector$new(c(1,2,3))
-#' popped <- pv$pop()
-#' pv$push(4)
-#' pv$push(c(5,6))
-#' print(pv) 
-#'
 #' @importFrom R6 R6Class
-#' @export
+#' 
+#' @keywords internal
 pvector <- R6Class("pvector", list(
+    #' @field values A vector of values
     values = NULL,
+
+    #' @description
+    #' Create a pvector
+    #' @param values A vector of values
+    #' @return A new pvector  
     initialize = function(values=c()) {
         self$values <- values
     },
-    print = function(...) {
+    #' @description
+    #' Print pvector
+    #' @return NULL  
+    print = function() {
         base::print(self$values)
         invisible(self)
     },
+    #' @description
+    #' Get length of pvector
+    #' @return An integer  
     length = function() {
         base::length(self$values)
     },
+    #' @description
+    #' Pop vector
+    #' @return Popped value 
     pop = function() {
         if (length(self$values) > 0) {
             popped.value <- self$values[1]
@@ -45,6 +37,10 @@ pvector <- R6Class("pvector", list(
             return(popped.value)   
         }
     },
+    #' @description
+    #' Push values
+    #' @param pushed.values A vector of values
+    #' @return NULL
     push = function(pushed.values) {
         self$values <- c(self$values, pushed.values)
     }